Moved more debug code to dedicated workspace
