+++ Assets/Scenes/Debugging/Hands/Hand_Debug.unity
@@ -123,11 +123,6 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
---- !u!1 &28528594 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 4171595189632753852, guid: 594ce1711b5217f438c8a8137bb9627a, type: 3}
-  m_PrefabInstance: {fileID: 357412575}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1 &31144542
 GameObject:
   m_ObjectHideFlags: 0
@@ -412,7 +407,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 275874967}
-  - component: {fileID: 275874976}
+  - component: {fileID: 275874978}
   - component: {fileID: 275874975}
   - component: {fileID: 275874970}
   - component: {fileID: 275874977}
@@ -515,55 +510,6 @@ MonoBehaviour:
   modelComplexity: 1
   refineFaceLandmarks: 1
   smoothLandmarks: 1
---- !u!114 &275874976
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 275874966}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3cfe681175a913848b21b465618cb95a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  debugTransforms:
-  - {fileID: 1000521036}
-  - {fileID: 54668406}
-  screen: {fileID: 1197972573}
-  graphRunner: {fileID: 275874975}
-  textureFramePool: {fileID: 275874970}
-  _worldAnnotationArea: {fileID: 1936307554}
-  _poseDetectionAnnotationController: {fileID: 462278167}
-  _holisticAnnotationController: {fileID: 462278166}
-  _poseWorldLandmarksAnnotationController: {fileID: 1936307555}
-  _poseRoiAnnotationController: {fileID: 462278164}
-  defaultVrmPrefab: {fileID: 4171595189632753852, guid: 594ce1711b5217f438c8a8137bb9627a, type: 3}
-  vrmModel: {fileID: 1321532676}
-  blendShapeProxy: {fileID: 1321532675}
-  animator: {fileID: 1321532674}
-  guiScript: {fileID: 1271193817}
-  sphereContainer: {fileID: 54668407}
-  keep: 0
-  TestInterpolation: 0
-  InterpolationValue: 0.1
-  TestA: 0
-  TestB: 0
-  TestC: 0
-  fixThumbTest: {x: 0, y: 0, z: 0}
-  fixHand: 1
-  handGroup: {fileID: 576270828}
-  handPositionOffset: {x: 0.1126, y: -0.06, z: 0.05}
-  handPositionScale: {x: -0.05, y: 0.05, z: 0.05}
-  rHandPosTest: {x: 0, y: 1.3, z: 0.5}
-  aaa: {x: 0, y: 0, z: 0}
-  bbb: {x: 0, y: 0, z: 0}
-  ccc: {x: 0, y: 0, z: 0}
-  ddd: {x: 0, y: -90, z: 90}
-  useCustomAnimZ: 0
-  useCustomAnimA: 0
-  useCustomAnimB: 0
-  useCustomAnimC: 0
 --- !u!114 &275874977
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -576,6 +522,42 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f3bb2c5c2d9ad6d44a4f7f5ae55782b3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &275874978
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 275874966}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f0ba77d0d46ea004fa71b270ee3fcf14, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  debugTransforms:
+  - {fileID: 1000521036}
+  - {fileID: 54668406}
+  screen: {fileID: 1197972573}
+  graphRunner: {fileID: 275874975}
+  textureFramePool: {fileID: 275874970}
+  defaultVrmPrefab: {fileID: 4171595189632753852, guid: 594ce1711b5217f438c8a8137bb9627a, type: 3}
+  vrmModel: {fileID: 1321532676}
+  blendShapeProxy: {fileID: 1321532675}
+  animator: {fileID: 1321532674}
+  guiScript: {fileID: 1271193817}
+  keep: 0
+  TestInterpolation: 0
+  InterpolationValue: 0.1
+  fixThumbTest: {x: 0, y: 0, z: 0}
+  fixHand: 0
+  handPositionOffset: {x: 0.1126, y: -0.06, z: 0.05}
+  handPositionScale: {x: -0.05, y: 0.05, z: 0.05}
+  handGroup: {fileID: 576270828}
+  _worldAnnotationArea: {fileID: 1252747460}
+  _poseDetectionAnnotationController: {fileID: 462278167}
+  _holisticAnnotationController: {fileID: 462278166}
+  _poseWorldLandmarksAnnotationController: {fileID: 1936307555}
+  _poseRoiAnnotationController: {fileID: 462278164}
 --- !u!1001 &283850634
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -806,7 +788,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4171595189632753852, guid: 594ce1711b5217f438c8a8137bb9627a, type: 3}
       propertyPath: m_IsActive
-      value: 1
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 4223638193890010088, guid: 594ce1711b5217f438c8a8137bb9627a, type: 3}
       propertyPath: m_RootOrder
@@ -2477,7 +2459,8 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   lineLength: 0.015
-  vrmModel: {fileID: 28528594}
+  alpha: 0.25
+  vrmModel: {fileID: 1321532676}
   rightHand: 1
   leftHand: 1
 --- !u!4 &1903942792
@@ -6404,7 +6387,7 @@ PrefabInstance:
     - target: {fileID: 9007664439247176511, guid: f9fef039014d09e47afc8086b765211f, type: 3}
       propertyPath: holisticSolution
       value: 
-      objectReference: {fileID: 275874976}
+      objectReference: {fileID: 0}
     - target: {fileID: 9007664439247176511, guid: f9fef039014d09e47afc8086b765211f, type: 3}
       propertyPath: modelTransform.z
       value: 0
@@ -6437,6 +6420,18 @@ PrefabInstance:
     - target: {fileID: 2193568136182861757, guid: 073f073358c23a643a9bfa14de4556c5, type: 3}
       propertyPath: m_RootOrder
       value: 10
+      objectReference: {fileID: 0}
+    - target: {fileID: 2193568136182861757, guid: 073f073358c23a643a9bfa14de4556c5, type: 3}
+      propertyPath: m_LocalScale.x
+      value: 0.01
+      objectReference: {fileID: 0}
+    - target: {fileID: 2193568136182861757, guid: 073f073358c23a643a9bfa14de4556c5, type: 3}
+      propertyPath: m_LocalScale.y
+      value: 0.01
+      objectReference: {fileID: 0}
+    - target: {fileID: 2193568136182861757, guid: 073f073358c23a643a9bfa14de4556c5, type: 3}
+      propertyPath: m_LocalScale.z
+      value: 0.01
       objectReference: {fileID: 0}
     - target: {fileID: 2193568136182861757, guid: 073f073358c23a643a9bfa14de4556c5, type: 3}
       propertyPath: m_LocalPosition.x
@@ -6477,6 +6472,10 @@ PrefabInstance:
     - target: {fileID: 2193568136182861757, guid: 073f073358c23a643a9bfa14de4556c5, type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3290018228251370116, guid: 073f073358c23a643a9bfa14de4556c5, type: 3}
+      propertyPath: scale
+      value: 0.002
       objectReference: {fileID: 0}
     - target: {fileID: 3852977236069312166, guid: 073f073358c23a643a9bfa14de4556c5, type: 3}
       propertyPath: m_Name
