+++ Assets/Scripts/Solution/HolisticSolutionBase.cs
@@ -1,16 +1,8 @@
-using Mediapipe;
 using Mediapipe.Unity;
 using System.Collections;
-using UnityEngine;
 
 namespace HardCoded.VRigUnity {
 	public abstract class HolisticSolutionBase : ImageSourceSolution<HolisticTrackingGraph> {
-		[SerializeField] protected RectTransform _worldAnnotationArea;
-		[SerializeField] protected DetectionAnnotationController _poseDetectionAnnotationController;
-		[SerializeField] protected HolisticLandmarkListAnnotationController _holisticAnnotationController;
-		[SerializeField] protected PoseWorldLandmarkListAnnotationController _poseWorldLandmarksAnnotationController;
-		[SerializeField] protected NormalizedRectAnnotationController _poseRoiAnnotationController;
-
 		// Always 'Full'
 		public HolisticTrackingGraph.ModelComplexity ModelComplexity {
 			get => graphRunner.modelComplexity;
@@ -37,11 +29,6 @@ namespace HardCoded.VRigUnity {
 		public float MinTrackingConfidence {
 			get => graphRunner.MinTrackingConfidence;
 			set => graphRunner.MinTrackingConfidence = value;
-		}
-
-		protected override void SetupScreen(ImageSource imageSource) {
-			base.SetupScreen(imageSource);
-			_worldAnnotationArea.localEulerAngles = imageSource.rotation.Reverse().GetEulerAngles();
 		}
 
 		protected abstract override void OnStartRun();
