+++ Assets/Scripts/Solution/HolisticTrackingSolutionWithDebug.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f0ba77d0d46ea004fa71b270ee3fcf14
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
