+++ Assets/Scripts/Utils/Groups/Interpolation.cs
@@ -1,0 +1,5 @@
+﻿using UnityEngine;
+
+namespace HardCoded.VRigUnity {
+	// TODO: Place interplotation structs here
+}
