+++ Assets/Scripts/Utils/Groups/Interpolation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0fd53d616ead30a4a9a9d909e12ca6ff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
